+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/PCMReader.kt
@@ -120,12 +120,7 @@ class PCMReader(
   private fun getAudioFormat(): Int = AudioFormat.ENCODING_PCM_16BIT
 
   private fun getChannelsConfig(): Int {
-    var numChannels = mediaFormat.getInteger(MediaFormat.KEY_CHANNEL_COUNT)
-
-    val channelCounts = config.device?.channelCounts
-    if (channelCounts != null && channelCounts.isNotEmpty()) {
-      numChannels = channelCounts.minBy { abs(it - numChannels) }
-    }
+    val numChannels = mediaFormat.getInteger(MediaFormat.KEY_CHANNEL_COUNT)
 
     return if (numChannels == 1) AudioFormat.CHANNEL_IN_MONO else AudioFormat.CHANNEL_IN_STEREO
   }