+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/format/AmrWbFormat.kt
@@ -13,7 +13,6 @@ class AmrWbFormat : Format() {
   )
 
   override val mimeTypeAudio: String = MediaFormat.MIMETYPE_AUDIO_AMR_WB
-  override val passthrough: Boolean = false
 
   override fun getMediaFormat(config: RecordConfig): MediaFormat {
     val format = MediaFormat().apply {