+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/format/FlacFormat.kt
@@ -11,7 +11,6 @@ class FlacFormat : Format() {
   )
 
   override val mimeTypeAudio: String = MediaFormat.MIMETYPE_AUDIO_FLAC
-  override val passthrough: Boolean = false
 
   override fun getMediaFormat(config: RecordConfig): MediaFormat {
     val format = MediaFormat().apply {